fix: move fixtures
