+++ tests/integration/hub_usage/test_hub_usage.py
@@ -1,6 +1,7 @@
 import os
 from pathlib import Path
 
+import docker
 import pytest
 from jina.logging import JinaLogger
 
@@ -242,6 +243,31 @@ def test_create_new(tmpdir, new_type):
     HubIO(args).new(no_input=True)
     list_dir = os.listdir(str(tmpdir))
     assert len(list_dir) == 1
+
+
+@pytest.fixture(scope='function')
+def test_workspace(tmpdir):
+    os.environ['JINA_TEST_JOINT'] = str(tmpdir)
+    workspace_path = os.environ['JINA_TEST_JOINT']
+    yield workspace_path
+    del os.environ['JINA_TEST_JOINT']
+
+
+@pytest.fixture(scope='function')
+def docker_image():
+    def _filter_repo_tag(image, image_name='jinahub/pod.dummy_mwu_encoder'):
+        tags = image.attrs['RepoTags'] if 'RepoTags' in image.attrs else None
+        if tags:
+            return tags[0].startswith(image_name)
+        else:
+            return False
+
+    img_name = 'jinahub/pod.dummy_mwu_encoder:0.0.6'
+    client = docker.from_env()
+    client.images.pull(img_name)
+    images = client.images.list()
+    image_name = list(filter(lambda image: _filter_repo_tag(image), images))[0]
+    return image_name
 
 
 def test_hub_build_level_pass(monkeypatch, test_workspace, docker_image):
